requirements.txt
numpy==2.3.2
requests==2.31.0
pandas==2.3.1
great_expectations==0.18.0
pymongo==4.4.1
python-dotenv==1.0.0
sendgrid==6.10.0
pytest==8.4.1
pytest-mock==3.14.1
flake8==7.3.0

